Extraction for blend files is now handled by a single class
